Rename stuff

+++ src/main/java/dev/ngspace/hudder/api/functionsandconsumers/HudderBuiltInMethods.java
@@ -4,18 +4,18 @@ import java.io.IOException;
 
 import dev.ngspace.hudder.Hudder;
 import dev.ngspace.hudder.api.compilers.interfaces.VariablesManager;
-import dev.ngspace.hudder.uielements.BuiltInTextureElement;
-import dev.ngspace.hudder.uielements.ColorVerticesElement;
-import dev.ngspace.hudder.uielements.GameHudElement;
-import dev.ngspace.hudder.uielements.GameHudElement.GuiType;
-import dev.ngspace.hudder.uielements.GradientElement;
-import dev.ngspace.hudder.uielements.ItemElement;
-import dev.ngspace.hudder.uielements.RectangleElement;
-import dev.ngspace.hudder.uielements.TextElement;
-import dev.ngspace.hudder.uielements.Texture9SliceElement;
-import dev.ngspace.hudder.uielements.TextureElement;
-import dev.ngspace.hudder.uielements.TextureVerticesElement;
+import dev.ngspace.hudder.uielements.minecraft.GameHudElement;
+import dev.ngspace.hudder.uielements.minecraft.ItemElement;
+import dev.ngspace.hudder.uielements.minecraft.GameHudElement.GuiType;
+import dev.ngspace.hudder.uielements.primitives.ColorVerticesElement;
+import dev.ngspace.hudder.uielements.primitives.GradientElement;
 import dev.ngspace.hudder.uielements.primitives.LineElement;
+import dev.ngspace.hudder.uielements.primitives.RectangleElement;
+import dev.ngspace.hudder.uielements.primitives.TextElement;
+import dev.ngspace.hudder.uielements.textures.BuiltInTextureElement;
+import dev.ngspace.hudder.uielements.textures.Texture9SliceElement;
+import dev.ngspace.hudder.uielements.textures.TextureElement;
+import dev.ngspace.hudder.uielements.textures.TextureVerticesElement;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
@@ -56,7 +56,7 @@ public class HudderBuiltInMethods {
 		
 		api.registerPositionedConsumer(
 				(e,_,_,_,s) -> e.addUIElement(
-						new TextureElement(s[0].asString(), s[1].asInt(), s[2].asInt(), s[3].asInt(), s[4].asInt())),
+						new TextureElement(s[0].asString(), s[1].asFloat(), s[2].asFloat(), s[3].asFloat(), s[4].asFloat())),
 				"drawLocalTexture", "drawPNG", "drawImage", "image", "png");
 		
 		api.registerPositionedConsumer((e,_,_,_,s) -> e.addUIElement(new Texture9SliceElement(s[0].asString(),
@@ -71,13 +71,13 @@ public class HudderBuiltInMethods {
 
 		api.registerPositionedConsumer(
 				(e,_,_,_,s) -> e.addUIElement(
-						new GradientElement(s[0].asInt(), s[1].asInt(), s[2].asInt(), s[3].asInt(),
+						new GradientElement(s[0].asFloat(), s[1].asFloat(), s[2].asFloat(), s[3].asFloat(),
 								s[4].asInt(), s[5].asInt(), false)),
 				"gradient");
 
 		api.registerPositionedConsumer(
 				(e,_,_,_,s) -> e.addUIElement(
-						new GradientElement(s[0].asInt(), s[1].asInt(), s[2].asInt(), s[3].asInt(),
+						new GradientElement(s[0].asFloat(), s[1].asFloat(), s[2].asFloat(), s[3].asFloat(),
 								s[4].asInt(), s[5].asInt(), true)),
 				"horizontal_gradient");
 		
@@ -108,17 +108,17 @@ public class HudderBuiltInMethods {
 		// GUI
 		
 		api.registerPositionedConsumer(
-				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asInt(), s[1].asInt(), GuiType.STATUS_BARS)),
+				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asFloat(), s[1].asFloat(), GuiType.STATUS_BARS)),
 				"drawStatusBars", "statusbars");
 		api.registerPositionedConsumer(
 				(e, _, _,_, s) -> e
-						.addUIElement(new GameHudElement(s[0].asInt(), s[1].asInt(), GuiType.EXP_AND_MOUNT_BAR)),
+						.addUIElement(new GameHudElement(s[0].asFloat(), s[1].asFloat(), GuiType.EXP_AND_MOUNT_BAR)),
 				"drawExpAndMountBars", "xpbar");
 		api.registerPositionedConsumer(
-				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asInt(), s[1].asInt(), GuiType.HOTBAR)),
+				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asFloat(), s[1].asFloat(), GuiType.HOTBAR)),
 				"drawHotbar", "hotbar");
 		api.registerPositionedConsumer(
-				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asInt(), s[1].asInt(), GuiType.ITEM_TOOLTIP)),
+				(e, _, _,_, s) -> e.addUIElement(new GameHudElement(s[0].asFloat(), s[1].asFloat(), GuiType.ITEM_TOOLTIP)),
 				"drawItemTooltip", "helditemtooltip");
 		
 		// Variables
@@ -138,58 +138,58 @@ public class HudderBuiltInMethods {
 		
 		// Items
 		
-		api.registerPositionedConsumer((e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(),
+		api.registerPositionedConsumer((e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(),
 				(mc.player.getVehicle() instanceof AbstractHorse horse) ? horse.getBodyArmorItem() : ItemStack.EMPTY,
 				s[2].asFloat(), false)), "drawMountArmor", "mountarmor");
 		
 		api.registerPositionedConsumer(
-				(e,_,_,c,s) -> e.addUIElement(new ItemElement(s[1].asInt(), s[2].asInt(),
+				(e,_,_,c,s) -> e.addUIElement(new ItemElement(s[1].asFloat(), s[2].asFloat(),
 						new ItemStack(BuiltInRegistries.ITEM.getValue(s[0].asIdentifier())), s.length > 3 ? s[3].asFloat() : c.scale(), false)),
 				"drawItem", "item");
 		
 		api.registerPositionedConsumer(
-				(e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(),
+				(e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(),
 								mc.player.getInventory().getItem(mc.player.getInventory().getSelectedSlot()),
 								s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"hand", "selectedslot");
 		
 		api.registerPositionedConsumer(
 				(e,_,_,_,s) -> e
-						.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(), mc.player.getInventory().getItem(39),
+						.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(), mc.player.getInventory().getItem(39),
 								s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"hat", "helmet");
 		
 		api.registerPositionedConsumer((e,_,_,_,s) -> e
-				.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(), mc.player.getInventory().getItem(38),
+				.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(), mc.player.getInventory().getItem(38),
 						s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"chestplate");
 		
 		api.registerPositionedConsumer(
 				(e,_,_,_,s) -> e
-						.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(), mc.player.getInventory().getItem(37),
+						.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(), mc.player.getInventory().getItem(37),
 								s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"leggings", "pants");
 		
 		api.registerPositionedConsumer((e,_,_,_,s) -> e
-				.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(), mc.player.getInventory().getItem(36),
+				.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(), mc.player.getInventory().getItem(36),
 						s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"boots");
 		
-		api.registerPositionedConsumer((e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asInt(), s[1].asInt(),
+		api.registerPositionedConsumer((e,_,_,_,s) -> e.addUIElement(new ItemElement(s[0].asFloat(), s[1].asFloat(),
 				mc.player.getOffhandItem(), s.length > 2 ? s[2].asFloat() : 1, s.length <= 3 || s[3].asBoolean())),
 				"offhand");
 		
 		// Slot
 		
 		api.registerPositionedConsumer(
-				(e,_,_,c,s) -> e.addUIElement(new ItemElement(s[1].asInt(), s[2].asInt(),
+				(e,_,_,c,s) -> e.addUIElement(new ItemElement(s[1].asFloat(), s[2].asFloat(),
 								mc.player.getInventory().getItem(s[0].asInt()),
 								s.length > 3 ? s[3].asFloat() : c.scale(), s.length <= 4 || s[4].asBoolean())),
 				"drawSlot", "slot");
 		
 		// Armor
 		api.registerPositionedConsumer(
-				(e,_,_,_,s) -> e.addUIElement(new ItemElement(s[1].asInt(), s[2].asInt(),
+				(e,_,_,_,s) -> e.addUIElement(new ItemElement(s[1].asFloat(), s[2].asFloat(),
 								mc.player.getInventory().getItem(36 + s[0].asInt()), s[3].asFloat(), s[4].asBoolean())),
 				"drawArmor", "armor");
